Debugged, all test are now valid, fixed weapon damage and the logic behind the calculations

+++ src/main/java/no/experis/dungeonmaster/characters/Wizard.java
@@ -36,7 +36,8 @@ public class Wizard extends Hero {
     }
 
     @Override
-    public int getDamage() {
-        return (int) (calculateDamage() * (1 + levelAttributes.getIntelligence() / 100.0));
+    public double getDamage() {
+
+        return (calculateDamage() * (1.00 + (levelAttributes.getIntelligence() / 100.00)));
     }
 }
